Address code review comment to move Region to a top level class
